[BE-20] 면접관 IDP 서버 동기화 및 관리

+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/interviewer/domain/InterviewerRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InterviewerRepository extends JpaRepository<Interviewer, Integer> {}
+public interface InterviewerRepository extends JpaRepository<Interviewer, Long> {}
